+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/modelo/PneuMarcaModeloResource.java
@@ -34,16 +34,21 @@ public final class PneuMarcaModeloResource {
     @GET
     @Secured(permissions = {Pilares.Frota.Pneu.CADASTRAR, Pilares.Frota.Pneu.ALTERAR, Pilares.Frota.Pneu.VISUALIZAR})
     @Path("/marcas")
+    @AppVersionCodeHandler(
+            targetVersionCode = 89,
+            versionCodeHandlerMode = VersionCodeHandlerMode.BLOCK_THIS_VERSION_AND_BELOW,
+            actionIfVersionNotPresent = VersionNotPresentAction.BLOCK_ANYWAY)
     public List<PneuMarcaListagem> getListagemMarcasPneu(
-            @QueryParam("codEmpresa") @Required Long codEmpresa,
-            @QueryParam("comModelos") @Optional boolean comModelos) {
-        return service.getListagemMarcasPneu(codEmpresa, comModelos);
+            @QueryParam("codEmpresa") @Required final Long codEmpresa,
+            @QueryParam("comModelos") @Optional final boolean comModelos,
+            @QueryParam("incluirMarcasNaoUtilizadas") @DefaultValue("true") final boolean incluirMarcasNaoUtilizadas) {
+        return service.getListagemMarcasPneu(codEmpresa, comModelos, incluirMarcasNaoUtilizadas);
     }
 
     @POST
     @Secured(permissions = {Pilares.Frota.Pneu.CADASTRAR, Pilares.Frota.Pneu.ALTERAR})
     @Path("/modelos")
-    public ResponseWithCod insertModeloPneu(@Valid final PneuModeloInsercao pneuModeloInsercao) {
+    public ResponseWithCod insertModeloPneu(@Required @Valid final PneuModeloInsercao pneuModeloInsercao) {
         return service.insertModeloPneu(pneuModeloInsercao);
     }
 
@@ -51,16 +56,15 @@ public final class PneuMarcaModeloResource {
     @Secured(permissions = {Pilares.Frota.Pneu.CADASTRAR, Pilares.Frota.Pneu.ALTERAR})
     @Path("/modelos")
     public ResponseWithCod updateModeloPneu(
-            @HeaderParam("Authorization") final String userToken,
-            @Valid final PneuModeloEdicao pneuModeloEdicao) {
+            @Required @Valid final PneuModeloEdicao pneuModeloEdicao) {
         return service.updateModeloPneu(pneuModeloEdicao);
     }
 
     @GET
     @Secured(permissions = {Pilares.Frota.Pneu.CADASTRAR, Pilares.Frota.Pneu.ALTERAR, Pilares.Frota.Pneu.VISUALIZAR})
     @Path("/modelos")
-    public List<PneuModeloListagem> getListagemModelosPneu(@QueryParam("codEmpresa") Long codEmpresa,
-                                                           @QueryParam("codMarca") Long codMarca) {
+    public List<PneuModeloListagem> getListagemModelosPneu(@QueryParam("codEmpresa") @Optional final Long codEmpresa,
+                                                           @QueryParam("codMarca") @Optional final Long codMarca) {
         return service.getListagemModelosPneu(codEmpresa, codMarca);
     }
 
@@ -71,7 +75,7 @@ public final class PneuMarcaModeloResource {
             targetVersionCode = 87,
             versionCodeHandlerMode = VersionCodeHandlerMode.BLOCK_THIS_VERSION_AND_BELOW,
             actionIfVersionNotPresent = VersionNotPresentAction.BLOCK_ANYWAY)
-    public PneuModeloVisualizacao getModeloPneu(@PathParam("codModelo") Long codModelo) {
+    public PneuModeloVisualizacao getModeloPneu(@PathParam("codModelo") @Required final Long codModelo) {
         return service.getModeloPneu(codModelo);
     }
 }