+++ src/main/java/br/com/zalf/prolog/webservice/implantacao/conferencia/frota/veiculo/VeiculoConferenciaService.java
@@ -2,11 +2,11 @@ package br.com.zalf.prolog.webservice.implantacao.conferencia.frota.veiculo;
 
 import br.com.zalf.prolog.webservice.Injection;
 import br.com.zalf.prolog.webservice.commons.gson.GsonUtils;
+import br.com.zalf.prolog.webservice.commons.network.Response;
 import br.com.zalf.prolog.webservice.commons.util.Log;
 import br.com.zalf.prolog.webservice.commons.util.date.Now;
 import br.com.zalf.prolog.webservice.errorhandling.exception.ProLogException;
-import br.com.zalf.prolog.webservice.implantacao.ImplantacaoImportTokensValidator;
-import br.com.zalf.prolog.webservice.implantacao.ImplantacaoImportTokens;
+import br.com.zalf.prolog.webservice.implantacao.autenticacao.ImplantacaoLoginSenhaValidator;
 import br.com.zalf.prolog.webservice.implantacao.conferencia.frota.veiculo.model.VeiculoPlanilha;
 import com.google.common.io.Files;
 import org.apache.commons.io.IOUtils;
@@ -27,19 +27,21 @@ public final class VeiculoConferenciaService {
     @NotNull
     private final VeiculoConferenciaDao dao = Injection.provideVeiculoConferenciaDao();
 
-    void getVerificacaoPlanilhaImportVeiculo(@NotNull final String token,
-                                             @NotNull final OutputStream out,
-                                             @NotNull final Long codUnidade,
-                                             @NotNull final InputStream fileInputStream,
-                                             @NotNull final FormDataContentDisposition fileDetail) {
+    public Response getVerificacaoPlanilhaImportVeiculo(@NotNull final String authorization,
+                                                        @NotNull final Long codEmpresa,
+                                                        @NotNull final Long codUnidade,
+                                                        @NotNull final InputStream fileInputStream,
+                                                        @NotNull final FormDataContentDisposition fileDetail) throws ProLogException {
         try {
-            ImplantacaoImportTokensValidator.validateTokenFor(ImplantacaoImportTokens.IMPORT_VEICULO, token);
-
+            final String usuario = new ImplantacaoLoginSenhaValidator().verifyUsernamePassword(authorization);
             final File file = createFileFromImport(codUnidade, fileInputStream, fileDetail);
-            readAndInsertImport(out, codUnidade, file);
+            readAndInsertImport(codEmpresa, codUnidade, usuario, file);
+            return Response.ok("Upload realizado com sucesso!");
         } catch (final Throwable throwable) {
             Log.e(TAG, "Erro ao verificar planilha de import de veiculos", throwable);
-            throw new RuntimeException(throwable);
+            throw Injection
+                    .provideProLogExceptionHandler()
+                    .map(throwable, "Erro com a conexão");
         }
     }
 
@@ -65,13 +67,14 @@ public final class VeiculoConferenciaService {
         }
     }
 
-    private void readAndInsertImport(@NotNull final OutputStream out,
+    private void readAndInsertImport(@NotNull final Long codEmpresa,
                                      @NotNull final Long codUnidade,
+                                     @NotNull final String usuario,
                                      @NotNull final File file) throws ProLogException {
         try {
             final List<VeiculoPlanilha> veiculoPlanilha = VeiculoPlanilhaReader.readListFromCsvFilePath(file);
             String jsonPlanilha = GsonUtils.getGson().toJson(veiculoPlanilha);
-            dao.getVerificacaoPlanilhaImportVeiculo(out, codUnidade, jsonPlanilha);
+            dao.importPlanilhaVeiculos(codEmpresa, codUnidade, usuario, jsonPlanilha);
         } catch (Throwable e) {
             Log.e(TAG, "Erro ao enviar dados para o BD", e);
             throw Injection