+++ src/main/java/br/com/zalf/prolog/webservice/implantacao/conferencia/frota/veiculo/VeiculoPlanilhaReader.java
@@ -33,7 +33,7 @@ public final class VeiculoPlanilhaReader {
         final String extension = FilenameUtils.getExtension(file.getName());
         if (extension.equalsIgnoreCase("xlsx")) {
             try {
-                new XlsxConverter().convertFileToCsv(file, 0, new SimpleDateFormat("ddMMyyyy"));
+                new XlsxConverter().convertFileToCsv(file, 2, new SimpleDateFormat("ddMMyyyy"));
             } catch (final IOException ex) {
                 throw new RuntimeException("Erro ao converter de XLSX para CSV", ex);
             }
@@ -41,7 +41,7 @@ public final class VeiculoPlanilhaReader {
         final CsvParserSettings settings = new CsvParserSettings();
         settings.setDelimiterDetectionEnabled(true, ',', ';');
         settings.setHeaderExtractionEnabled(true);
-        settings.setNumberOfRowsToSkip(0);
+        settings.setNumberOfRowsToSkip(12);
         final CsvParser parser = new CsvParser(settings);
         final List<String[]> rows = parser.parseAll(file);
         final List<VeiculoPlanilha> veiculoPlanilha = new ArrayList<>();
@@ -55,37 +55,33 @@ public final class VeiculoPlanilhaReader {
     }
 
     private static VeiculoPlanilha read(@NotNull final String[] linha) {
-        if (linha[0].isEmpty()) {
+        if (linha[1].isEmpty()) {
             return null;
         }
         final VeiculoPlanilha item = new VeiculoPlanilha();
-
         // PLACA.
-        item.setPlaca(linha[0]);
-
-        // KM.
         if (!StringUtils.isNullOrEmpty(linha[1])) {
-            item.setKm(Long.parseLong(linha[1]));
+            item.setPlaca(linha[1]);
         }
-
         // MARCA.
         if (!StringUtils.isNullOrEmpty(linha[2])) {
             item.setMarca(linha[2]);
         }
-
         // MODELO.
         if (!StringUtils.isNullOrEmpty(linha[3])) {
             item.setModelo(linha[3]);
         }
-
-        // TIPO.
+        // KM.
         if (!StringUtils.isNullOrEmpty(linha[4])) {
-            item.setTipo(linha[4]);
+            item.setKm(Long.parseLong(linha[4]));
         }
-
-        // DIAGRAMA.
+        // TIPO.
         if (!StringUtils.isNullOrEmpty(linha[5])) {
-            item.setDiagrama(linha[5]);
+            item.setTipo(linha[5]);
+        }
+        // DIAGRAMA.
+        if (!StringUtils.isNullOrEmpty(linha[6])) {
+            item.setQtdEixos(linha[6]);
         }
         return item;
     }