+++ src/main/java/br/com/zalf/prolog/webservice/integracao/operacoes/OperacoesIntegradasMovimentacao.java
@@ -4,6 +4,8 @@ import br.com.zalf.prolog.webservice.frota.pneu.movimentacao._model.ProcessoMovi
 import br.com.zalf.prolog.webservice.frota.pneu.servico.ServicoDao;
 import org.jetbrains.annotations.NotNull;
 
+import java.time.LocalDateTime;
+
 /**
  * Created on 14/10/19.
  *
@@ -13,5 +15,6 @@ public interface OperacoesIntegradasMovimentacao {
     @NotNull
     Long insert(@NotNull final ServicoDao servicoDao,
                 @NotNull final ProcessoMovimentacao processoMovimentacao,
+                @NotNull final LocalDateTime dataHoraMovimentacao,
                 final boolean fecharServicosAutomaticamente) throws Throwable;
 }